Clean up macro a bit
